refactor(device): Cameras → Domain/Models (+10 erreurs strict corrigées en code), Fleet contrats morts supprimés (implémente #6842, audit 2026-09-05)
