+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+- **refactor(device): Cameras — modèles Eloquent déplacés vers Domain/Models/ + 10 erreurs PHPStan strict corrigées ; Fleet — contrats morts supprimés (implémentation #6842, audit 2026-09-05, PR #6832).** (1) `api/app/Modules/Cameras/Domain/{Camera,CameraAccessLog,CameraAccessToken,CameraPermission}.php` → `Domain/Models/` (namespace `App\Modules\Cameras\Domain\Models`), 15 fichiers consommateurs mis à jour (controllers, services, Policies, AuthServiceProvider, tests). (2) 10 erreurs PHPStan strict (niveau 8) résolues en code : trait `HasFactory` retirée des 3 modèles (aucune factory ni usage `::factory()` dans le repo), `fresh() ?? $instance` sur les retours de `CameraService` (create/update/issueAccessToken/revokeAccessToken), accès null-safe `$token->camera?->company`, garde explicite `camera_id NOT NULL` dans `CameraService::log()` (throw `InvalidArgumentException` — chemin auparavant en erreur DB), `present($permission->fresh() ?? $permission)` dans `CameraPermissionController`. Vérifié : PHPStan modules (niveau 5) 0 erreur, PHPStan strict ciblé 0 erreur, Pint OK, baseline strict **inchangée** (cliquets anti-dette #5448/PA2-ARCH-005 préservés — la réduction comptable se fera en PR dédiée). (3) `Modules/Fleet/Domain/Contracts/{Trip,Vehicle}RepositoryInterface` supprimées (0 référence, 0 implémentation — grep repo entier) ; `docs/ARCHITECTURE_STATUS.md` ligne Fleet alignée (Contracts → —). Aucun changement de comportement API.
 - **chore(governance): audit 2026-09-05, lot 3 — registre BC purgé des chemins `planned` fantômes, arbre de décision complété, backlog matérialisé (issues #6841/#6842).** (1) `bounded-context-registry.json` : chemins `api/app/Solutions/FuelStation|EduManager` (`planned`, organisation Solutions supprimée) retirés des BC-15/16 actifs ; BC-20 Documents conservé (module futur documenté — DEP_BC20). Garde MAT-001 verte. (2) `api/ARCHITECTURE.md` : arbre de décision complété d'un tableau des zones transverses réelles (`app/Http/{Controllers/Web,Resources,Middleware}`, `app/Jobs`, `app/Console`, `app/Policies`, `app/Contracts`, `app/Events`, `app/Shared/Models`, `app/AI`) avec propriétaires/BC — règle rappelée : tout code métier va dans `Modules/<BC>/`. (3) Backlog : issue #6841 (inventaire mesuré des couches `Application/` vides, à décliner par BC) + #6842 (Cameras → `Domain/Models/`). Aucun changement de code applicatif.
 - **docs(contrib): chemins de fichiers explicités dans CONTRIBUTING.md et GUIDELINES.md (`docs/plan/PLAN_100PCT.md`, `docs/GESTION_PROJET/SCENARIOS_TEST_API_GITHUB_ACTIONS.md`) — audit PM 2026-09-05.**
 - **docs(cleanup): plan d'action audit 2026-09-05, lot 2 — fichier PHP corrompu restauré, route-owners complété, CHANGELOG re-structuré (branche unifiée, PR #6832).** (1) `api/app/Modules/HR/Application/Actions/ApplySectorTemplate.php` : fichier corrompu par les merges union (sans `<?php`/namespace depuis 5509530e1) restauré depuis son ancêtre lint-propre `5113abf67` (106 l., classe valide, 0 référence — précédent purge #6817/#6818). (2) `dev-hub/governance/route-owners.json` : 6 fichiers de routes post-consolidation déclarés (travelagency `mixed`, solutions `mixed`, restaurantmanager/fuel_station/edu_manager/delivery `tenant`) — 29/29 fichiers, garde MAT-003 couche statique verte. (3) `CHANGELOG.md` : 10 bulletins orphelins placés au-dessus du header `## [Unreleased]` (résidus de merges union) rapatriés sous le header ; fragments « ⚠️ En attente de merge » collés en fin de 2 bulletins et blocquote orphelin supprimés. (4) Comptages « 18 modules » → 25 dans `docs/architecture/COMPTABILITE_CONCEPTION.md` et `docs/CONTEXT/02_TECHNICAL_CONTEXT.md`. (5) `AGENTS.md` : périmètre du registre BC mis à jour (BC-01 … BC-26 DELIVERY). Aucun changement de comportement applicatif.
