+++ api/app/Modules/Cameras/Application/Actions/DeleteCamera.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace App\Modules\Cameras\Application\Actions;
 
 use App\Core\Auth\Domain\Models\Employee;
-use App\Modules\Cameras\Domain\Camera;
+use App\Modules\Cameras\Domain\Models\Camera;
 use App\Modules\Cameras\Infrastructure\Services\CameraService;
 
 class DeleteCamera
