+++ api/app/Modules/Cameras/Interfaces/Api/V1/Controllers/CameraController.php
@@ -10,8 +10,8 @@ use App\Http\Controllers\Controller;
 use App\Modules\Cameras\Application\Actions\CreateCamera;
 use App\Modules\Cameras\Application\Actions\DeleteCamera;
 use App\Modules\Cameras\Application\Actions\UpdateCamera;
-use App\Modules\Cameras\Domain\Camera;
-use App\Modules\Cameras\Domain\CameraPermission;
+use App\Modules\Cameras\Domain\Models\Camera;
+use App\Modules\Cameras\Domain\Models\CameraPermission;
 use App\Modules\Cameras\Infrastructure\Services\CameraService;
 use App\Modules\Cameras\Interfaces\Api\V1\Requests\StoreCameraRequest;
 use App\Modules\Cameras\Interfaces\Api\V1\Requests\TestRtspRequest;
