+++ api/app/Modules/Cameras/Interfaces/Api/V1/Controllers/PublicCameraViewerController.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace App\Modules\Cameras\Interfaces\Api\V1\Controllers;
 
 use App\Http\Controllers\Controller;
-use App\Modules\Cameras\Domain\Camera;
-use App\Modules\Cameras\Domain\CameraAccessToken;
+use App\Modules\Cameras\Domain\Models\Camera;
+use App\Modules\Cameras\Domain\Models\CameraAccessToken;
 use App\Modules\Cameras\Infrastructure\Services\CameraService;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
