+++ docs/ARCHITECTURE_STATUS.md
@@ -17,7 +17,7 @@
 | **EdgeSync** | ✅ | — | — | ✅ | — | ✅ | ✅ | ✅ | ✅ |
 | **EduManager** | ✅ | ✅ | ✅ | ✅ | — | ✅ | ✅ | ✅ | ✅ |
 | **Expense** | ✅ | — | ✅ | — | — | ✅ | ✅ | ✅ | ✅ |
-| **Fleet** | ✅ | ✅ | ✅ | — | — | — | ✅ | ✅ | ⚠️ |
+| **Fleet** | ✅ | — | ✅ | — | — | — | ✅ | ✅ | ⚠️ |
 | **FuelStation** | ✅ | ✅ | ✅ | — | — | ✅ | ✅ | ✅ | ✅ |
 | **Growth** | ✅ | ✅ | ✅ | ✅ | ✅ | — | ✅ | ✅ | ✅ |
 | **HR** | ✅ | ✅ | ✅ | ✅ | ✅ | ✅ | ✅ | ✅ | ✅ |
